pyproject.toml
[project]
name = "cex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "tomli>=2.4.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]
